requirements.txt
Django==2.1.3
djangorestframework==3.9.0
Faker==0.9.2
pkg-resources==0.0.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
text-unidecode==1.2
